chore(docs): clean up exposed symbols and doc comments for Rust SDK
