+++ sdk/rust/Cargo.toml
@@ -32,3 +32,7 @@ console = "0.15.10"
 macros = ["dep:hipcheck-sdk-macros"]
 mock_engine = []
 print-timings = []
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
