+++ sdk/rust/src/error.rs
@@ -108,6 +108,7 @@ impl From<Infallible> for Error {
 	}
 }
 
+/// A Result type using `hipcheck_sdk::Error`
 pub type Result<T> = StdResult<T, Error>;
 
 /// Errors specific to the execution of `Plugin::set_configuration()` to configure a Hipcheck
